Store and score functionality!
Introduces basic store functionality.

Main changes:
+Added shopScreen to prefabs.
+Added ScoreManager to prefabs.
+Added ShopManager to prefabs.
+Added a new font
+Added the textmeshpro system, which allows for dynamic text changes and 3d-powered rendering of fonts (text)
+Added a new folder to separate music and sound effects in the "Audiofiles" folder
+Added an error sound effect
+Added a "ca-chinggg" sound effect
--------------------------------------------------
moneyAndScoreManager.cs:
Tracks the current score and money of the player.

Avaliable members:
-int: currentMoney
-int: prevMoney
-int: currentScore

Debug features:
    - Press "k" to give yourself $100.
    - Press "b" to simulate an object being purchased for $450
    - Press "h" to set the amount of money avaliable to $450

IMPORTANT: Scores and money are both represented by INTEGERS to facilitate easy subtraction and addition, to remove rounding errors.

+giveMoney(int giveAmount, bool doublePoints)
Gives the amount of money to the player. If the latter bool is true the amount is multiplied by two. Perfect for powerup shenanigans.

+removeMoney(int cost)
Removes money from the players wallet.

+giveScore(int score)
Gives score to the player. SCORE IS NOT TO BE CONFUSED WITH MONEY
--------------------------------------------------
ShopScript.cs:
+updateShopSigns()
Updates the shop signs and dynamically changes progress bars on each item for easy visual identification of how close you are to being able to purchase an item

The script goes through each child of the "Products" child of the canvas. It finds the text field "price" and casts this value to an int.
It calculates the float adequateMoneyRatio to fill a rounded rectangle dynamically. Im sure theres a better way to do this but this seems to be both performant and stable.

+purchaseObject()
(WIP) purchase an object, removes the money from the user and plays a "ca-ching" sound if the funds are avaliable
Else it goes "err-eeerrrr", which is a little funny.

------------------------------------
The shop itself:
- Built using a "Canvas" 2D renderer named shopScreen
     - The canvas has 5 children:
        - AudioSource object - keeps track of where store-related sounds are played from
        - SHOPTITLE - Displays the "Kessler Foundation Store". What is the kessler foundation? idk is there even lore in this game lol
        - moneyCounter - Displays the users current money
        - Products - Contains all the avaliable products for purchasing. Each Product child is divided like this:
            - Text(TMP) - Displays the name of the product
            - price - Displays and contains the price of the product. This is where you edit the price if you wish. No arrays, only children,
            - dynBorder - Displays and contains the dynamic border.
       - Plane - The background material of the store.

Im sure we could convert the store items to an array/matrix of names and prices, but this is the way i got it to work, and as stated before it seems to be performant enough using only the main thread.