Added Statistics screen in Gameover + Player info tracking
The game (playerStatTracker.cs) tracks main user metrics each second (time interval and array size customizable) and presents them when the game is over (WindowGraph.cs). Currently the stats screen isn't in VR and doesn't scale to amount of game ticks.

The stats screen is red when there was a wave at that point in time.
Currently the stats screen tracks:
+ Enemies killed.
+ User Health.
+ User Score
+ User Money.
+ Cooldown or not that tick.